+++ tests/Test Cases/mix.txt
@@ -1,0 +1,1 @@
+Hello! How are you? I am great. 2 + 2 = 4 @£$%^&*(){}[];:'"\|,./<>?`~§± this is a test case with mixed characters
