+++ tests/Test Cases/repeated_char.txt
@@ -1,0 +1,1 @@
+aaaaaaaaaaaaaaaabbbbbbbbbbbbbbbccccccccccccccccdddddddddddddddeeeeeeeeeeeeeee
