+++ tests/Test Cases/simple_phrase.txt
@@ -1,0 +1,1 @@
+This is a simple phrase for a test case
